Move data from template to view

+++ places/migrations/0003_auto_20250411_1842.py
@@ -1,4 +1,5 @@
 from places.models import Place, Image
+from django.db.models import Q
 from django.db import migrations
 import requests
 import codecs
@@ -25,22 +26,29 @@ def download_image_and_get_path_and_name(img_url, place_name, img_id):
         os.makedirs(directory)
     response = requests.get(img_url)
     response.raise_for_status()
-    with open(path_to_img, 'wb') as file:
-        file.write(response.content)
+    if not os.path.exists(path_to_img):
+        with open(path_to_img, 'wb') as file:
+            file.write(response.content)
     return path_to_img, img_name
     
+
+def update_place_relationships(place_name):
+    place = Place.objects.get(Q(title__contains=place_name))
+    place.images.set(Image.objects.filter(Q(title__contains=place_name)))
+
 
 def parse_images(apps, schema_editor):
     path = 'static/places/'
     for file_id, filename in enumerate(os.listdir(path)):
-        places = get_images_and_name_of_place(path, filename)
-        for img_number, img in enumerate(places['images']):
-            path_to_img, img_name = download_image_and_get_path_and_name(img, places['name'], img_number+1)
+        place = get_images_and_name_of_place(path, filename)
+        for img_number, img in enumerate(place['images']):
+            path_to_img, img_name = download_image_and_get_path_and_name(img, place['name'], img_number+1)
             Image.objects.create(
-                place = Place.objects.get(id = file_id),
+                # place = Place.objects.get(id = file_id),
                 title = img_name,
                 image = path_to_img
             )
+        update_place_relationships(place['name'])
 
 
 def move_backward(apps, schema_editor):
@@ -51,7 +59,7 @@ def move_backward(apps, schema_editor):
 class Migration(migrations.Migration):
 
     dependencies = [
-        ('places', '0003_image'),
+        ('places', '0002_auto_20250410_0908'),
     ]
 
     operations = [
